fixed some broken tests from lower int casts on some parameters and added in an initial sketch of boundary conditions, at least for dirichlet bcs.

+++ src/quadrature_templates/QuadQuadrature.jl
@@ -1,4 +1,4 @@
-function quad_quadrature_points_init!(quadrature_order::Int,
+function quad_quadrature_points_init!(quadrature_order::Int8,
                                       w::Array{Float64,1},
                                       ξ::Array{Float64,2})
     if quadrature_order == 1
